requirements.txt
fastecdsa==2.2.3
jsonschema==3.2.0
matplotlib==3.5.2
numpy==1.21.5
pycryptodome==3.15.0
rsa==4.9
scipy==1.7.3
